Refactor HeroSectionAdminView to use Bootstrap 5
Moved HeroSectionAdminView to a new module under home/hero_view.py with updated templates for Bootstrap 5. Updated admin route registration to use the new view and added a category. Cleaned up imports and removed the old Bootstrap 4 implementation from site_views.py.
