+++ app/routes/admin/routes.py
@@ -17,8 +17,10 @@ from .views import (
 # === View Registration ===
 admin.add_view(ParentAdminView(Parent, db.session))
 admin.add_view(PuppyAdminView(Puppy, db.session))
+# === Updated  Bootstap 5 View Registration === #
 admin.add_view(ReviewAdminView(Review, db.session)) # <-- bootstap 5 LINE
-admin.add_view(HeroSectionAdminView(HeroSection, db.session, name="Hero Section"))
+#admin.add_view(HeroSectionAdminView(HeroSection, db.session, name="Hero Section"))
+admin.add_view(HeroSectionAdminView(HeroSection, db.session, name="Hero Section", category="Home"))
 admin.add_view(AboutSectionAdminView(AboutSection, db.session, name="About Section"))
 admin.add_view(AnnouncementBannerAdminView(AnnouncementBanner, db.session, name="Announcement Banner"))
 admin.add_view(AdminModelView(GalleryImage, db.session, name="Gallery Images"))
