+++ app/routes/admin/views/__init__.py
@@ -3,5 +3,7 @@
 from .base import MyAdminIndexView, AdminModelView
 from .parent_views import ParentAdminView
 from .puppy_views import PuppyAdminView
-from .site_views import HeroSectionAdminView, AboutSectionAdminView, AnnouncementBannerAdminView
+from .site_views import AboutSectionAdminView, AnnouncementBannerAdminView
 from .home.review_view import ReviewAdminView
+from .home.hero_view import HeroSectionAdminView 
+
